Extract parser.py from legacy.py
